OXDEV-940 Move Container to Application package
