+++ source/Application/Controller/AccountController.php
@@ -5,7 +5,7 @@
  */
 namespace OxidEsales\EshopCommunity\Application\Controller;
 
-use OxidEsales\EshopCommunity\Internal\Common\Container\Container;
+use OxidEsales\EshopCommunity\Internal\Application\Container;
 
 /**
  * Current user "My account" window.
